# docs/04-api-contract.yaml
openapi: 3.0.3
info:
  title: My-AI API
  version: 0.1.0
  description: |
    V1.1 接口契约（知识库主数据化 + 文档列表管理基线）

    RAG 权限体系阶段二开始后，除登录接口和健康检查外，`/api/v1/**` 默认需要
    `HttpOnly Cookie` 中的服务端 Session。所有写操作需要携带固定 Header：
    `X-MYAI-CSRF: 1`。
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/admin/knowledge-bases/{kbId}/grants:
    get:
      tags: [governance]
      summary: 查询知识库 ACTIVE 授权列表
      description: |
        返回当前工作区内指定知识库的 ACTIVE 授权列表；
        当前列表口径固定为：
        - `knowledge_base_grants.status = ACTIVE`
        - 被授权用户 `users.status = ACTIVE`
        - 被授权成员关系 `workspace_memberships.status = ACTIVE`
        仅工作区所有者或管理员可访问。
      x-stage: implemented
      security:
        - cookieAuth: []
      parameters:
        - name: kbId
          in: path
          required: true
          schema:
            type: string
          description: 知识库业务键 kb_id
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KnowledgeBaseGrant'
        '400':
          description: 路径参数非法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 当前用户无知识库授权治理权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 知识库不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-frontend-note: |
        前端知识库授权管理页采用“整表勾选 + 保存”交互，但保存时按成员差量调用
        `PUT /api/v1/admin/knowledge-bases/{kbId}/grants/{userId}` 与
        `DELETE /api/v1/admin/knowledge-bases/{kbId}/grants/{userId}`；
        当前资源维度不提供 `grants:batch` 接口。
  /api/v1/admin/knowledge-bases/{kbId}/grants/{userId}:
    put:
      tags: [governance]
      summary: 授予或更新知识库授权
      description: |
        为当前工作区内指定知识库授予或更新单个成员的知识库角色；
        授权对象必须是当前工作区有效成员；
        同一 `(workspace_id, kb_id, user_id)` 授权关系保持单一真源；
        授权变更后应立即影响既有知识库、文档与问答访问控制结果。
      x-stage: implemented
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
        - name: kbId
          in: path
          required: true
          schema:
            type: string
          description: 知识库业务键 kb_id
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: 目标成员用户 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertKnowledgeBaseGrantRequest'
      responses:
        '200':
          description: 授权成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeBaseGrant'
        '400':
          description: 请求体为空、路径参数非法或知识库角色值非法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 缺少或非法 CSRF Header，或当前用户无知识库授权治理权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 知识库不存在，或目标用户不是当前工作区有效成员
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [governance]
      summary: 回收知识库授权
      description: |
        回收当前工作区内指定知识库对单个成员的 ACTIVE 授权；
        回收成功后应立即影响既有知识库、文档与问答访问控制结果。
      x-stage: implemented
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
        - name: kbId
          in: path
          required: true
          schema:
            type: string
          description: 知识库业务键 kb_id
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: 目标成员用户 ID
      responses:
        '204':
          description: 回收成功
        '400':
          description: 路径参数非法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 缺少或非法 CSRF Header，或当前用户无知识库授权治理权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 知识库不存在或 ACTIVE 授权不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/documents/{documentId}/grants:
    get:
      tags: [governance]
      summary: 查询文档 ACTIVE 授权列表
      description: |
        返回当前工作区内指定文档的 ACTIVE 授权列表；
        当前列表口径固定为：
        - `document_grants.status = ACTIVE`
        - 被授权用户 `users.status = ACTIVE`
        - 被授权成员关系 `workspace_memberships.status = ACTIVE`
        仅工作区所有者或管理员可访问。
      x-stage: implemented
      security:
        - cookieAuth: []
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
          description: 文档资产 ID
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentGrant'
        '400':
          description: 路径参数非法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 当前用户无文档授权治理权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 文档不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-frontend-note: |
        前端文档授权管理页采用“整表勾选 + 保存”交互，但保存时按成员差量调用
        `PUT /api/v1/admin/documents/{documentId}/grants/{userId}` 与
        `DELETE /api/v1/admin/documents/{documentId}/grants/{userId}`；
        当前资源维度不提供 `grants:batch` 接口。
  /api/v1/admin/documents/{documentId}/grants/{userId}:
    put:
      tags: [governance]
      summary: 授予或更新文档授权
      description: |
        为当前工作区内指定文档授予或更新单个成员的文档级权限覆盖；
        授权对象必须是当前工作区有效成员；
        同一 `(workspace_id, document_id, user_id)` 授权关系保持单一真源；
        文档级覆盖继续遵循 `DOC_DENY` 最高优先级。
      x-stage: implemented
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
        - name: documentId
          in: path
          required: true
          schema:
            type: string
          description: 文档资产 ID
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: 目标成员用户 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertDocumentGrantRequest'
      responses:
        '200':
          description: 授权成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentGrant'
        '400':
          description: 请求体为空、路径参数非法或文档权限值非法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 缺少或非法 CSRF Header，或当前用户无文档授权治理权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 文档不存在，或目标用户不是当前工作区有效成员
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [governance]
      summary: 回收文档授权
      description: |
        回收当前工作区内指定文档对单个成员的 ACTIVE 授权；
        回收成功后应立即影响既有文档列表、文档读取与问答过滤结果。
      x-stage: implemented
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
        - name: documentId
          in: path
          required: true
          schema:
            type: string
          description: 文档资产 ID
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: 目标成员用户 ID
      responses:
        '204':
          description: 回收成功
        '400':
          description: 路径参数非法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 缺少或非法 CSRF Header，或当前用户无文档授权治理权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 文档不存在或 ACTIVE 授权不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/admin/members/{userId}/document-grants:
    get:
      tags: [governance]
      summary: 查询成员维度文档授权列表
      description: |
        返回指定有效成员在当前工作区下的文档 ACTIVE 授权列表；
        仅工作区所有者或管理员可访问。
      x-stage: implemented
      security:
        - cookieAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: 目标成员用户 ID
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentGrant'
        '400':
          description: 路径参数非法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 当前用户无成员治理权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 目标成员不存在或不是当前工作区有效成员
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/members/{userId}/document-grants:batch:
    put:
      tags: [governance]
      summary: 批量覆盖成员维度文档授权
      description: |
        以“成员”为维度整体替换其文档授权集合；
        不在请求列表中的现有 ACTIVE 授权将被回收；
        仅工作区所有者或管理员可访问。
      x-stage: implemented
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: 目标成员用户 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                assignments:
                  type: array
                  items:
                    type: object
                    required: [documentId, permission]
                    properties:
                      documentId:
                        type: string
                      permission:
                        type: string
                        enum: [DOC_ALLOW_READ, DOC_ALLOW_MANAGE, DOC_DENY]
      responses:
        '200':
          description: 覆盖成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentGrant'
        '400':
          description: 请求体为空、路径参数非法或权限值非法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 缺少或非法 CSRF Header，或当前用户无成员治理权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 目标成员不存在或不是当前工作区有效成员
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/documents:
    get:
      tags: [ingest]
      summary: 查询文档分页列表
      description: |
        返回文档管理台所需的分页结果；
        默认按 `createdAt DESC` 排序；
        当未传 `status` 时，默认排除 `DELETED`；
        查询范围限定在当前用户所属工作区，并对结果按知识库授权与文档级覆盖做可读过滤。
      x-stage: implemented
      security:
        - cookieAuth: []
      parameters:
        - name: kbId
          in: query
          required: false
          schema:
            type: string
          description: 知识库业务键 kb_id；传入不存在值时返回空列表而非 400
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [UPLOADED, INGESTING, INDEXED, FAILED, DELETING, DELETED]
          description: 不传时默认排除 DELETED；显式传 DELETED 时仅查询已删除文档
        - name: filename
          in: query
          required: false
          schema:
            type: string
            maxLength: 200
          description: 文件名模糊匹配
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: 每页条数
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
          description: 分页偏移量
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListPageResponse'
        '400':
          description: 参数非法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 当前用户无文档读取权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/documents/{documentId}:
    delete:
      tags: [ingest]
      summary: 删除文档资产
      x-stage: implemented
      description: |
        软删除文档资产，清理源文件与该 documentId 的全部向量版本；
        删除成功后可通过 status 接口查询到 DELETED；
        当前用户必须对目标文档具备管理权限。
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
        - name: documentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 删除成功（或已删除，幂等成功）
        '404':
          description: 文档不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 缺少或非法 CSRF Header，或当前用户无目标文档管理权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 文档状态冲突（INGESTING/DELETING）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 删除执行失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  parameters:
    CsrfHeader:
      name: X-MYAI-CSRF
      in: header
      required: true
      schema:
        type: string
        enum: ['1']
      description: 写操作统一 CSRF Header，当前固定值为 `1`
  schemas:
...
    KnowledgeBaseGrant:
      type: object
      properties:
        workspaceId:
          type: string
          example: default
        kbId:
          type: string
        userId:
          type: string
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [KB_MANAGER, KB_CONTRIBUTOR, KB_READER, KB_ASKER]
        status:
          type: string
          enum: [ACTIVE]
    UpsertKnowledgeBaseGrantRequest:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [KB_MANAGER, KB_CONTRIBUTOR, KB_READER, KB_ASKER]
    DocumentGrant:
      type: object
      properties:
        workspaceId:
          type: string
          example: default
        documentId:
          type: string
        userId:
          type: string
        username:
          type: string
        displayName:
          type: string
        permission:
          type: string
          enum: [DOC_ALLOW_READ, DOC_ALLOW_MANAGE, DOC_DENY]
        status:
          type: string
          enum: [ACTIVE]
    UpsertDocumentGrantRequest:
      type: object
      required: [permission]
      properties:
        permission:
          type: string
          enum: [DOC_ALLOW_READ, DOC_ALLOW_MANAGE, DOC_DENY]
...
    DocumentListPageResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DocumentListItemResponse'
        total:
          type: integer
          format: int64
        limit:
          type: integer
        offset:
          type: integer
    DocumentListItemResponse:
      type: object
      properties:
        documentId:
          type: string
        kbId:
          type: string
        filename:
          type: string
        fileSize:
          type: integer
          format: int64
        status:
          type: string
          enum: [UPLOADED, INGESTING, INDEXED, FAILED, DELETING, DELETED]
        failureReason:
          type: string
          nullable: true
          description: 仅 FAILED 状态时返回实际失败原因，其余状态返回 null
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...